submits/itertools

See merge request 2023-se/se-226-2-Nikita-Oleynik-oleynik_n_a!54

+++ tasks/itertools/test.cpp
@@ -61,7 +61,7 @@ TEST_CASE("Zip short") {
         ++it;
     }
     REQUIRE(it == s.end());
-    REQUIRE(i == s.size());
+    REQUIRE(static_cast<size_t>(i) == s.size());
 }
 
 struct Int {
